requirements.txt
pyyaml
schematics